first api call

+++ app/src/main/java/com/example/shopapp/data/network/model/Rating.kt
@@ -2,5 +2,5 @@ package com.example.shopapp.data.network.model
 
 data class Rating(
     val count: Int,
-    val rate: Int
+    val rate: Float
 )
